requirements.txt
langchain
langchain-groq
streamlit
pypdf
faiss-cpu
openai
typing-extensions
langgraph
python-dotenv
scikit-learn
numpy
transformers
langchain_community
fastapi
uvicorn
python-multipart
pydantic
starlette
uuid